+++ doc/CHANGES.md
@@ -7,6 +7,9 @@
 
 * Even though it's only necessary for the results of dynamic queries, everything is decorated right now.
 
+## 0.3.0 - Eliminating Brute Force
+* First attempt at eliminating the brute force array scanning approach of the past versions.
+
 ## 0.2.0 - First Stable Release
 * The library is stable enough now for a general release.